PythonAIAgentFromScratch-main/PythonAIAgentFromScratch-main/requirements.txt
langchain-mistralai
langchain
langchain-community
python-dotenv
pydantic
wikipedia
google-api-python-client
streamlit
reportlab